requirements.txt
ftfy
regex
tqdm
fast_pytorch_kmeans
torch~=2.2.1
pillow~=10.2.0